# backend/openapi.json
{
  "components": {
    "schemas": {
      "AdminExportJobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminExportJobRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminExportJobListResponse",
        "type": "object"
      },
...
      "AdminPresentationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminPresentationRow"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminPresentationListResponse",
        "type": "object"
      },
...
      "AdminSetupRead": {
        "properties": {
          "entra_client_id_configured": {
            "title": "Entra Client Id Configured",
            "type": "boolean"
          },
          "entra_client_secret_configured": {
            "title": "Entra Client Secret Configured",
            "type": "boolean"
          },
          "entra_enabled": {
            "title": "Entra Enabled",
            "type": "boolean"
          },
          "entra_redirect_uri": {
            "title": "Entra Redirect Uri",
            "type": "string"
          },
          "entra_tenant_id_configured": {
            "title": "Entra Tenant Id Configured",
            "type": "boolean"
          },
          "local_password_auth_enabled": {
            "title": "Local Password Auth Enabled",
            "type": "boolean"
          },
          "public_api_url": {
            "title": "Public Api Url",
            "type": "string"
          },
          "public_app_url": {
            "title": "Public App Url",
            "type": "string"
          }
        },
        "required": [
          "local_password_auth_enabled",
          "entra_enabled",
          "entra_tenant_id_configured",
          "entra_client_id_configured",
          "entra_client_secret_configured",
          "public_app_url",
          "public_api_url",
          "entra_redirect_uri"
        ],
        "title": "AdminSetupRead",
        "type": "object"
      },
      "AdminStatsRead": {
        "properties": {
          "app_log_rows_24h": {
            "title": "App Log Rows 24H",
            "type": "integer"
          },
          "audit_events_24h": {
            "title": "Audit Events 24H",
            "type": "integer"
          },
          "export_jobs": {
            "title": "Export Jobs",
            "type": "integer"
          },
          "presentations": {
            "title": "Presentations",
            "type": "integer"
          },
          "users": {
            "title": "Users",
            "type": "integer"
          },
          "versions": {
            "title": "Versions",
            "type": "integer"
          }
        },
        "required": [
          "users",
          "presentations",
          "versions",
          "export_jobs",
          "audit_events_24h",
          "app_log_rows_24h"
        ],
        "title": "AdminStatsRead",
        "type": "object"
      },
      "AdminUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AdminUserListResponse",
        "type": "object"
      },
...
      "AppLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AppLogRead"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "AppLogListResponse",
        "type": "object"
      },
...
      "AuditLogListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AuditLogListResponse",
        "type": "object"
      },
      "AuditLogRead": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "target_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Kind"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "id",
          "ts",
          "actor_id",
          "action",
          "target_kind",
          "target_id",
          "metadata",
          "ip"
        ],
        "title": "AuditLogRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresentationAccess": {
        "enum": [
          "admin",
          "owner",
          "editor",
          "user"
        ],
        "title": "PresentationAccess",
        "type": "string"
      },
      "PresentationCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "PresentationCreate",
        "type": "object"
      },
      "PresentationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PresentationRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PresentationListResponse",
        "type": "object"
      },
...
      "PresentationRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_user_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresentationAccess"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_id": {
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "title",
          "description",
          "current_version_id",
          "created_at",
          "updated_at"
        ],
        "title": "PresentationRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Internal presentation collaboration API (v1 scaffold).",
    "title": "promptDeck API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/audit": {
      "get": {
        "operationId": "list_audit_api_v1_admin_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Audit",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/jobs": {
      "get": {
        "operationId": "list_export_jobs_admin_api_v1_admin_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminExportJobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Export Jobs Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/logs": {
      "get": {
        "operationId": "list_logs_api_v1_admin_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "description": "Filter by channel (http, auth, audit, script)",
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by channel (http, auth, audit, script)",
              "title": "Channel"
            }
          },
          {
            "in": "query",
            "name": "request_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Request Id"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "description": "Only rows whose path starts with this prefix (e.g. /api/v1/admin)",
            "in": "query",
            "name": "path_prefix",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only rows whose path starts with this prefix (e.g. /api/v1/admin)",
              "title": "Path Prefix"
            }
          },
          {
            "description": "Substring match on `event` (e.g. auth.login, http.request)",
            "in": "query",
            "name": "event_contains",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on `event` (e.g. auth.login, http.request)",
              "title": "Event Contains"
            }
          },
          {
            "description": "Use `id` of oldest row from prior page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Use `id` of oldest row from prior page",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppLogListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Logs",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/presentations": {
      "get": {
        "operationId": "list_presentations_admin_api_v1_admin_presentations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPresentationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Presentations Admin",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/setup": {
      "get": {
        "operationId": "admin_setup_api_v1_admin_setup_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSetupRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Setup",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "operationId": "admin_stats_api_v1_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/presentations": {
      "get": {
        "operationId": "list_presentations_api_v1_presentations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Presentations",
        "tags": [
          "presentations"
        ]
      },
      "post": {
        "operationId": "create_presentation_api_v1_presentations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresentationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Presentation",
        "tags": [
          "presentations"
        ]
      }
    },
...
  }
}